fix(): added babel and library compilation
